while loop + some others
I added a boolean for the main method so that we could remove saying menu in each case, it still works by exiting if 7, but now returns to menu as part of the while loop, instead of saying menu in each case.

I also fixed my storeStudentScores, changing the name to inputStudentScores and also changing student scores array size to be as large as we declare it when we call upon the method.